Improve speed of comparing test bitmaps by an order of magnitude.
This compares the bitmaps by comparing their actual byte contents, instead
of retrieving individual pixel colors.

Also, the diff bitmap will now contain differences highlighted in red,
overlaid on top of the reference image.

The "percent" difference parameter shall now be a float, since the
difference can be smaller than 1% but larger than 0%. Previously this
would have been rounded down to 0, which is incorrect.

+++ app/src/androidTest/java/org/wikipedia/espresso/Constants.java
@@ -5,7 +5,7 @@ public final class Constants {
     public static final String TEST_OUTPUT_FOLDER = "/wikipedia-app-tests/";
     public static final String TEST_ASSET_FOLDER = "espresso/";
     public static final int SCREENSHOT_COMPRESSION_QUALITY = 100;
-    public static final int SCREENSHOT_COMPARE_PERCENT_TOLERANCE = 0;
+    public static final float SCREENSHOT_COMPARE_PERCENT_TOLERANCE = 0f;
 
     private Constants() { }
 }
